Add support for recursive assemblies (group instances within group instances). Fix missing root path in filelinks if file did exist already.
Also prevents hidden objects to show up in hybrid mode. (Using 'atom'
to prevent decomposition of assemblies still works.)

Currently in hybrid mode, assemblies are listed in the total bill of
materials (as well as all the parts and their total part counts that make up all the assemblies).

=> The object part counts are absolute and contain all required parts.
The assembly listing is just a help.

This will be especially useful for blue print and assembly instruction
manual generation. (But these are separate addons that just share the
decomposition algorithm.)