Support to include blueprints. Vary by volume.
Support to include blueprints as images. Blueprints are generated for
each object and may be useful for construction purposes, especially
that each volume variant of an object (depending on modifiers, e.g.
Boolean for holes, cuts, ...) gets its own engineering drawing.

Vary by volume: The extension keeps tracks of volume variants, i.e.
not only raw parts (e.g. plates of steel) are listed, but also the
special/postprocessed parts, e.g. how many of that kind with so many
holes and cuts (because this influences the volume).